+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+Fix issue when starting multiple time the same service
+
 ## v1.3.0
 
 - Add `filter: (key: string, data: Object) => boolean` on `whenEvent` and `whenResult`.